auth-microservice/requirements.txt
boto3
chalice
pyrebase4